🐛 [bug fix] fix bug in favorites
- use user.db for favorites
- extract list of words from user.db
